Refactor into subfiles
Refactor into subfiles for clarity
