+++ TeleportEverything/app.config
@@ -1,19 +1,19 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Mono.Cecil" publicKeyToken="50cebf1cceb9d05e" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-0.10.4.0" newVersion="0.10.4.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Mono.Cecil.Mdb" publicKeyToken="50cebf1cceb9d05e" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-0.10.4.0" newVersion="0.10.4.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Mono.Cecil.Pdb" publicKeyToken="50cebf1cceb9d05e" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-0.10.4.0" newVersion="0.10.4.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
+    <runtime>
+        <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+            <dependentAssembly>
+                <assemblyIdentity name="Mono.Cecil" publicKeyToken="50cebf1cceb9d05e" culture="neutral"/>
+                <bindingRedirect oldVersion="0.0.0.0-0.10.4.0" newVersion="0.10.4.0"/>
+            </dependentAssembly>
+            <dependentAssembly>
+                <assemblyIdentity name="Mono.Cecil.Mdb" publicKeyToken="50cebf1cceb9d05e" culture="neutral"/>
+                <bindingRedirect oldVersion="0.0.0.0-0.10.4.0" newVersion="0.10.4.0"/>
+            </dependentAssembly>
+            <dependentAssembly>
+                <assemblyIdentity name="Mono.Cecil.Pdb" publicKeyToken="50cebf1cceb9d05e" culture="neutral"/>
+                <bindingRedirect oldVersion="0.0.0.0-0.10.4.0" newVersion="0.10.4.0"/>
+            </dependentAssembly>
+        </assemblyBinding>
+    </runtime>
 </configuration>
