+++ TeleportEverything/packages.config
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="HarmonyX" version="2.5.5" targetFramework="net462" />
-  <package id="JotunnLib" version="2.4.7" targetFramework="net462" />
-  <package id="Mono.Cecil" version="0.11.4" targetFramework="net462" />
-  <package id="MonoMod" version="21.9.19.1" targetFramework="net462" />
-  <package id="MonoMod.RuntimeDetour" version="21.8.19.1" targetFramework="net462" />
-  <package id="MonoMod.Utils" version="21.9.19.1" targetFramework="net462" />
+    <package id="HarmonyX" version="2.5.5" targetFramework="net462"/>
+    <package id="JotunnLib" version="2.4.7" targetFramework="net462"/>
+    <package id="Mono.Cecil" version="0.11.4" targetFramework="net462"/>
+    <package id="MonoMod" version="21.9.19.1" targetFramework="net462"/>
+    <package id="MonoMod.RuntimeDetour" version="21.8.19.1" targetFramework="net462"/>
+    <package id="MonoMod.Utils" version="21.9.19.1" targetFramework="net462"/>
 </packages>
